requirements.txt
gdown-5.2.0
rembg
tqdm
prodigyopt
ptflops
wandb
torchinfo
